Show every three hours forecast

+++ app/src/main/java/io/github/a0gajun/weather/data/entity/owm_common/Main.java
@@ -25,7 +25,7 @@ public class Main {
 
     @Getter
     @SerializedName("pressure")
-    private int pressure;
+    private float pressure;
 
     @Getter
     @SerializedName("humidity")
