+++ app/src/main/java/io/github/a0gajun/weather/data/entity/owm_common/Rain.java
@@ -23,5 +23,5 @@ public class Rain {
      */
     @Getter
     @SerializedName("3h")
-    private int threeHour;
+    private float threeHour;
 }
