+++ app/src/main/java/io/github/a0gajun/weather/data/entity/owm_common/Snow.java
@@ -22,5 +22,5 @@ public class Snow {
      */
     @Getter
     @SerializedName("3h")
-    private int threeHour;
+    private float threeHour;
 }
