+++ app/src/main/java/io/github/a0gajun/weather/data/entity/owm_common/Wind.java
@@ -24,5 +24,5 @@ public class Wind {
 
     @Getter
     @SerializedName("deg")
-    private int deg;
+    private float deg;
 }
