+++ app/src/main/java/io/github/a0gajun/weather/domain/model/FiveDayForecast.java
@@ -43,7 +43,7 @@ public class FiveDayForecast {
         // main
         @Getter @Setter private float temperature;
 
-        @Getter @Setter private int pressure;
+        @Getter @Setter private float pressure;
 
         @Getter @Setter private int humidity;
 
@@ -69,10 +69,10 @@ public class FiveDayForecast {
         @Getter @Setter private int cloudiness;
 
         // rain
-        @Getter @Setter private int rainOfLast3Hour;
+        @Getter @Setter private float rainOfLast3Hour;
 
         // ic_snow
-        @Getter @Setter private int snowOfLast3Hour;
+        @Getter @Setter private float snowOfLast3Hour;
 
         /**
          * Data/time of caluclation
